+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changes
 
+**v0.6**
+   - reflecting Webex rebrand: CLU renamed `guestissuer`, deprecating 'sparkguest'
+   - changing '--spark' option to `guestissuer verify --access`
+   - upgraded dependencies
+
 **v0.5.1**
    - making the duration of the JWT Token more explicit
    - added explanations to add a Permanent Guest to an existing space